refactored arrows pages. Moved from parent to child
