+++ src/Routes/Root/index.js
@@ -3,70 +3,21 @@ import {
     StyledNavLink, StyledSearch, StyledBar,
     StyledMoviesBrowser, StyledNav, StyledIcon, StyledInput,
     StyledSection,
-    StyledPages,
-    StyledButton
 } from "./styled";
 import camera from "../../Images/camera.svg";
 import { Outlet, useLocation, Form, useSubmit, useNavigate } from "react-router-dom"
 import { searchQueryParamName } from "../../features/useQueryParameter";
 import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { axiosMovieList, selectMovieList } from "../../features/moveList/movieListSlice";
-import { useState } from "react";
-import leftArrow from "../../Images/leftArrow.svg";
-import rightArrow from "../../Images/rightArrow.svg"
-
-const firstPage = (navigate) => {
-    navigate(`/movies/page/1`)
-}
-
-const lastPage = (navigate) => {
-    navigate(`/movies/page/500`)
-}
-
-const addPage = (getPage, navigate) => {
-    navigate(`/movies/page/${(+getPage + 1)}`)
-}
-
-const subPage = (getPage, navigate) => {
-    navigate(`/movies/page/${(+getPage - 1)}`)
-}
-
-const getPopularMoviesPages = (getPage, navigate) => {
-    let disabledFirst = false;
-    let disabledLast = false
-
-    if(getPage < 2 ) {
-        disabledFirst = true;
-    }
-
-    if(getPage > 499 ) {
-        disabledLast = true;
-    }
-
-    return (
-        <StyledPages>
-            <StyledButton disabled={disabledFirst} onClick={() => firstPage(navigate)}> <img src={leftArrow}/> First</StyledButton>
-            <StyledButton disabled={disabledFirst} onClick={() => subPage(getPage, navigate)}> <img src={leftArrow}/> Previous</StyledButton>
-            {/* Due to talk TMDB, error occure over > 500 pages so cant load above that */}
-            <div>Page {getPage} of 500</div>
-            <StyledButton disabled={disabledLast} onClick={() => addPage(getPage, navigate)}>Next <img src={rightArrow}/> </StyledButton>
-            <StyledButton disabled={disabledLast} onClick={() => lastPage(navigate)}>Last <img src={rightArrow}/> </StyledButton>
-        </StyledPages>
-    )
-}
+import { useDispatch } from "react-redux";
+import { axiosMovieList } from "../../features/moveList/movieListSlice";
 
 const Root = () => {
-    const [getPage, setGetPages] = useState(1)
-
-    const getMovies = useSelector(selectMovieList);
     const location = useLocation();
-    const splitLocation = location.pathname.split('/');
     const submit = useSubmit();
     const navigate = useNavigate();
-    
+    const dispatch = useDispatch();
+
     const searchMovie = (new URLSearchParams(location.search)).get(searchQueryParamName) || null;
-    const URLPage = splitLocation[3];
 
     const onInputChange = ({ currentTarget }) => {
         if (currentTarget.value.trim() === "") {
@@ -76,15 +27,9 @@ const Root = () => {
         submit(currentTarget.form)
     };
 
-    const dispatch = useDispatch();
-
     useEffect(() => {
         dispatch(axiosMovieList());
-    }, [searchMovie])
-
-    useEffect(() => {
-        setGetPages(previousPage => previousPage = URLPage);
-    }, [URLPage])
+    }, [searchMovie, dispatch])
 
     return (
         <>
@@ -119,9 +64,8 @@ const Root = () => {
                 </StyledSearch>
             </StyledHeader>
             <StyledSection>
-                <Outlet context={getPage} />
+                <Outlet />
             </StyledSection>
-            {splitLocation[1] === "movies" && getMovies !== null ? getPopularMoviesPages(getPage, navigate) : ""}
         </>
     )
 };
