+++ src/features/moveList/ArrowPages/styled.js
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const StyledPages = styled.div`
+    display: flex;
+    gap:12px;
+    justify-content: center;
+    align-items: center;
+    margin: 40px 0 103px 0;
+`;
+
+export const StyledButton = styled.button`
+    cursor: pointer;
+    display: flex;
+    gap: 8px;
+    align-items: center;
+    background-color: #D6E4FF;
+    padding: 8px 16px;
+    border-radius: 5px;
+    border: none;
+`
